Modernizing the plugin
It seems that nowdays, the standard way to load plugins on logstash
is creating a gem, which is later installed on the machine. The make
creates the jar and the gem file. Right now, the jar is downloaded from
a maven repository at the moment of the intallation of the plugin. This
probably need to be revised to make the plugin more self contained.